requirements_combined.txt

#
# 
#   conda activate brazilian_sales
#   conda create -n brazilian_sales python=3.10
#   pip freeze > requirements-frozen.txt
#   pip install --upgrade -r requirements.txt
#   pip install -r requirements.txt
# ============================================================================
# API interactions
# API requests
# Brazilian Sales Analytics - Complete Requirements
# CLOUD & DATABASE
# Code quality
# Core Dagster
# CORE DATA TOOLS
# Create virtual environment:
# Dagster integrations
# dagster-slack==0.21.11
# Dashboard
# Data Engineering Stack: BigQuery + dbt + Python
# Data export
# Data manipulation
# Data manipulation and analysis
# DATA PROCESSING
# Date handling
# dbt - Data transformation
# dbt (data build tool) for BigQuery
# DEVELOPMENT & TESTING
# Email (optional)
# Environment and configuration
# Environment variables
# Freeze current environment:
# Google Cloud BigQuery
# Google Cloud Platform
# Great Expectations
# Install all dependencies: pip install -r requirements.txt
# Installation:
# JSON handling
# Jupyter notebooks
# Jupyter notebooks (optional)
# Kaggle
# Last Updated: 2025-12-08
# Logging and monitoring
# Meltano - Data integration (if using)
# Meltano and Singer ecosystem
# NOTES
# OPTIONAL: ALERTS & MONITORING
# ORCHESTRATION - DAGSTER
# Plotting
# Progress bars
# Python Dependencies
# sendgrid==6.10.0
# Slack notifications (optional)
# Streamlit
# Testing
# Type stubs
# Update all packages:
# UTILITIES
# Version: 2.0.0
# Visualization
# VISUALIZATION - STREAMLIT
# YAML parsing
bandit==1.9.2  # Security linting
black==23.12.1
black==25.12.0
coverage[toml]==7.12.0  # Coverage reporting
dagster-dbt==0.21.11
dagster-gcp==0.21.11
dagster-pandas==0.21.11
dagster-webserver==1.5.11
dagster==1.5.11
db-dtypes==1.4.4  # BigQuery data types for pandas
dbt-bigquery==1.10.3
dbt-core==1.10.15
flake8==7.0.0
flake8==7.3.0
google-auth==2.23.0
google-auth==2.43.0
google-cloud-bigquery==3.11.4
google-cloud-bigquery==3.38.0
google-cloud-secret-manager==2.25.0
google-cloud-storage==2.10.0
google-cloud-storage==3.1.0
great_expectations==1.9.2
ipykernel==7.1.0
ipython==8.18.1
isort==7.0.0  # Import sorting
jupyter==1.0.0
jupyter==1.1.1
kaggle==1.8.2
matplotlib==3.10.7
meltano==3.3.0
meltano==4.0.6
mypy==1.19.0
notebook==7.5.0
numpy==1.24.3
numpy==2.3.5
openpyxl==3.1.5  # Excel export
pandas-stubs==2.3.3.251201
pandas==2.0.3
pandas==2.3.3
plotly==5.18.0
plotly==6.5.0
pre-commit==4.5.0
pyarrow==22.0.0  # Fast columnar processing
pytest-cov==4.1.0
pytest-cov==7.0.0
pytest-mock==3.15.1  # Mocking helpers
pytest-timeout==2.4.0  # Test timeout
pytest-xdist==3.8.0  # Parallel test execution
pytest==7.4.3
pytest==9.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-dotenv==1.2.1
PyYAML==6.0.1
requests==2.31.0
requests==2.32.5
safety==3.7.0  # Updated for compatibility with packaging>=22.0
seaborn==0.13.2
streamlit-aggrid==1.2.1  # Enhanced dataframe display
streamlit==1.29.0
streamlit==1.52.1  # Updated for better compatibility
structlog==25.5.0
tqdm==4.66.1
tqdm==4.67.1
types-requests==2.32.4.20250913
ujson==5.9.0
